refactor: 💡 home 改为 example
