+++ template/src/store/index.ts
@@ -3,8 +3,8 @@ import {
   createLogger,
 } from 'vuex';
 import { importAllStore } from '@/libs/utils/importAllStore';
-import { Store as HomeStore } from '@/views/home/store';
-import { StateType as HomeState } from '@/views/home/store/state';
+import { Store as ExampleStore } from '@/views/example/store';
+import { StateType as ExampleState } from '@/views/example/store/state';
 import { UserState } from './modules/user';
 import { LogsState } from './modules/logs';
 import { MenusState } from './modules/menus';
@@ -16,14 +16,14 @@ export interface RootState {
 }
 
 export interface AllStateTypes extends RootState {
-  home: HomeState;
+  example: ExampleState;
   user: UserState;
   logs: LogsState;
   menus: MenusState;
   screenfull: ScreenfullState;
 }
 
-export type Store = HomeStore<Pick<AllStateTypes, 'home'>>
+export type Store = ExampleStore<Pick<AllStateTypes, 'example'>>
 
 const debug = process.env.NODE_ENV !== 'production';
 const plugins = debug ? [createLogger({})] : [];
