+++ template/src/views/example/store/index.ts
