+++ template/src/views/example/store/mutations.ts
