+++ template/src/views/example/store/state.ts
